refactor: give filter files different names to reduce confusion
