+++ site/src/pages/AuditPage/AuditFilter.tsx
@@ -11,7 +11,7 @@ import {
 	MenuSkeleton,
 	SearchFieldSkeleton,
 	type useFilter,
-} from "components/Filter/filter";
+} from "components/Filter/Filter";
 import {
 	type UseFilterMenuOptions,
 	useFilterMenu,
